// components/Checkbox/Checkbox.d.ts
import { default as React } from '../../../node_modules/react';
/**
 * Checkbox component for boolean user input.
 */
export declare const Checkbox: React.ForwardRefExoticComponent<React.InputHTMLAttributes<HTMLInputElement> & {
    label?: string;
    indeterminate?: boolean;
    className?: string;
} & React.RefAttributes<HTMLInputElement>>;